Add files via upload
Programmed Attiny 85 to read input from mechanical switches, along with a debouncing algorithm on a 5ms timer interrupt, to get input for safe code. If one were to press the switches in the correct order, the LEDS return to an unlocked/programming mode, and a new combination of 6 presses can be put in as our safe code. If the input combination of switches is wrong, we blink the yellow LED a few times and return to our locked mode, waiting for the right input to be submitted.